refactor: split drag drop selection tests
